Refactor TSS into structure that's more consistent with conventions
