+++ rest/data/select.go
@@ -1,0 +1,56 @@
+package data
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/evergreen-ci/evergreen"
+	"github.com/evergreen-ci/evergreen/rest/model"
+	testselection "github.com/evergreen-ci/test-selection-client"
+	"github.com/evergreen-ci/utility"
+	"github.com/pkg/errors"
+)
+
+// newTestSelectionClient constructs a new test selection client using the
+// provided HTTP client.
+func newTestSelectionClient(c *http.Client) *testselection.APIClient {
+	tssBaseURL := evergreen.GetEnvironment().Settings().TestSelection.URL
+	conf := testselection.NewConfiguration()
+	conf.HTTPClient = c
+	conf.Servers = testselection.ServerConfigurations{
+		testselection.ServerConfiguration{
+			URL:         tssBaseURL,
+			Description: "Test selection service",
+		},
+	}
+	return testselection.NewAPIClient(conf)
+}
+
+// SelectTests uses the test selection service to return a filtered set of tests
+// to run based on the provided SelectTestsRequest. It returns the list of
+// selected tests.
+func SelectTests(ctx context.Context, req model.SelectTestsRequest) ([]string, error) {
+	httpClient := utility.GetHTTPClient()
+	defer utility.PutHTTPClient(httpClient)
+
+	c := newTestSelectionClient(httpClient)
+	var strategies []testselection.StrategyEnum
+	for _, s := range req.Strategies {
+		strategies = append(strategies, testselection.StrategyEnum(s))
+	}
+	reqBody := testselection.BodySelectTestsApiTestSelectionSelectTestsProjectIdRequesterBuildVariantNameTaskIdTaskNamePost{
+		TestNames:  req.Tests,
+		Strategies: strategies,
+	}
+	selectedTests, resp, err := c.TestSelectionAPI.SelectTestsApiTestSelectionSelectTestsProjectIdRequesterBuildVariantNameTaskIdTaskNamePost(ctx, req.Project, req.Requester, req.BuildVariant, req.TaskID, req.TaskName).
+		BodySelectTestsApiTestSelectionSelectTestsProjectIdRequesterBuildVariantNameTaskIdTaskNamePost(reqBody).
+		Execute()
+	if resp != nil {
+		defer resp.Body.Close()
+	}
+	if err != nil {
+		return nil, errors.Wrap(err, "forwarding request to test selection service")
+	}
+	return selectedTests, nil
+}
+
